Cleaner encapsulation of the existing Jolt Test machinery.
next step is to break it down into ECS components and build the stuff needed to expose it to UE safely.
