+++ Source/Barrage/Public/BarrageDispatch.h
@@ -17,4 +17,5 @@
 		virtual void Initialize(FSubsystemCollectionBase& Collection) override;
 		virtual void OnWorldBeginPlay(UWorld& InWorld) override;
 		virtual void Deinitialize() override;
+		TStatId GetStatId() const override;
 	};
